rootfs: Use fbset when possible to set a specific mode.

+++ rootfs.nix
@@ -116,21 +116,28 @@ let
     }
 
     set_framebuffer_mode() {
+      if [ -e /etc/fb.modes ]; then
+        # Uses the first defined mode
+        fbset $(grep ^mode /etc/fb.modes | head -n1 | cut -d'"' -f2)
+      else
         [ -e "/sys/class/graphics/fb0/modes" ] || return
         [ -z "$(cat /sys/class/graphics/fb0/mode)" ] || return
 
         _mode="$(cat /sys/class/graphics/fb0/modes)"
         echo "Setting framebuffer mode to: $_mode"
         echo "$_mode" > /sys/class/graphics/fb0/mode
+      fi
     }
 
     echo 1 > /sys/class/leds/lcd-backlight/brightness
 
     set_framebuffer_mode
 
-    msm-fb-refresher --loop &
-
     show_splash loading
+
+    set_framebuffer_mode
+
+    msm-fb-refresher --loop &
 
     ln -sv ${shell} /bin/sh
 
@@ -258,6 +265,8 @@ let
       { object = stage1; symlink = "/init"; }
       { object = ./temp-splash.png; symlink = "/splash.png"; }
       { object = ./loading.png; symlink = "/loading.png"; }
+      # FIXME : configurable rootfs additions from device config.
+      { object = ./devices/asus-z00t/fb.modes; symlink = "/etc/fb.modes"; }
     ];
   };
 in
